GPII-1551:  Updated to newer infusion and newer conventions for gpii-express, directory paths, etc.
