+++ package.json
@@ -5,27 +5,25 @@
   "repository": "https://github.com/GPII/gpii-pouchdb",
   "license": "BSD-3-Clause",
   "scripts": {
-    "test": "node tests/js/all-tests.js",
-    "postinstall": "grunt dedupe-infusion"
+    "test": "node tests/js/all-tests.js"
   },
   "dependencies": {
     "express-pouchdb": "0.14.0",
-    "gpii-express": "git://github.com/the-t-in-rtf/gpii-express#ffc96bad77180b31f34654f268597caa93570683",
+    "gpii-express": "git://github.com/the-t-in-rtf/gpii-express#d7b995bc5d210439699c269ce8ef327280167550",
     "express": "4.13.3",
-    "grunt": "~0.4.4",
-    "grunt-shell": "0.6.4",
-    "grunt-contrib-jshint": "~0.9.0",
-    "grunt-jsonlint": "1.0.4",
-    "grunt-gpii": "git://github.com/GPII/grunt-gpii.git#58ae055833f437a9e91baedb9f39a22c5fdd727f",
-    "infusion": "git://github.com/fluid-project/infusion.git#282f1a318718eed0b0ec060fb8b4ad254417fd7e",
+    "infusion": "2.0.0-dev.20151211T025229Z.eabb4fc",
     "memdown": "1.0.0",
     "pouchdb": "3.3.1",
     "underscore-node": "0.1.2",
     "when": "^3.7.3"
   },
   "devDependencies": {
-    "jqUnit":     "git://github.com/fluid-project/node-jqUnit.git#b7ef53aac1ca365bf23d35d255fae874faa88b70",
-    "kettle":     "git://github.com/amb26/kettle.git#33e2826e78315c6d29b13c8ba810f54fef29e8ac",
+    "grunt": "~0.4.4",
+    "grunt-shell": "0.6.4",
+    "grunt-contrib-jshint": "~0.9.0",
+    "grunt-jsonlint": "1.0.4",
+    "kettle": "git://github.com/amb26/kettle.git#2517cc1d9d047d58aef70db48d1580da718ffc16",
+    "node-jqunit": "~1.1.3",
     "request": "2.51.0",
     "tough-cookie": "0.12.1"
   }
