+++ tests/js/all-tests.js
@@ -1,6 +1,6 @@
 // Run all tests in this package in series.
 "use strict";
-var fluid = fluid || require("infusion");
+var fluid = require("infusion");
 
 require("./basic-tests");
 require("./reload-tests");
