+++ tests/js/launch-test-harness.js
@@ -1,6 +1,6 @@
 // A convenience script to start the test harness, used for manual QA.
 "use strict";
-var fluid = fluid || require("infusion");
+var fluid = require("infusion");
 var gpii  = fluid.registerNamespace("gpii");
 
 require("./harness");
