--- tests/lib/sequence.js
@@ -1,21 +1,0 @@
-// Common function to prepend common sequences to all test modules in a testCaseHolder.
-"use strict";
-var fluid = fluid || require("infusion");
-var gpii  = fluid.registerNamespace("gpii");
-
-// Wire in a specified set of starting sequences to a set of tests.
-fluid.registerNamespace("gpii.pouch.tests");
-gpii.pouch.tests.addRequiredSequences = function (sequenceStart, rawTests) {
-    var completeTests = fluid.copy(rawTests);
-
-    for (var a = 0; a < completeTests.length; a++) {
-        var testSuite = completeTests[a];
-        for (var b = 0; b < testSuite.tests.length; b++) {
-            var tests = testSuite.tests[b];
-            var modifiedSequence = sequenceStart.concat(tests.sequence);
-            tests.sequence = modifiedSequence;
-        }
-    }
-
-    return completeTests;
-};
